Remove PXLBLCK_HIGHER_COLOR_RESOLUTION_ENABLED option
The option PXLBLCK_HIGHER_COLOR_RESOLUTION_ENABLED was added in the early days of the pxlBlck. It should give an option to switch between 24 and 16Bit color mode. Since there is no (real) advantage at using the 16Bit color mode this option was now deactivated and the 24Bit color mode is enabled by default.